// src/types/custom-d-t.ts
export interface FAQItemDT {
  id: string;
  question: string;
  answer: string;
  isOpen?: boolean;
}